Update SN Ia module
